feat: add secret mode
If your function does not want to be called publicly, then use the secret parameter. This will generate a token for you, which the user sets via a web page or URL, and the authenticated user can only call the function, otherwise only the parameters can be previewed.

Also, I removed feat from `37599d3` because using gunicorn and gevent in funix could cause multi-threaded errors that could probably be fixed in the future.